clean up functions and upload working notebook
